fix ternary overflow issue

+++ src/requirements.cpp
@@ -464,7 +464,7 @@ std::vector<std::string> requirement_data::get_folded_list( int width,
                 component.has( crafting_inv, filter, batch ) ) {
                 const itype_id item_id = static_cast<itype_id>( component.type );
                 const long qty = item::count_by_charges( item_id ) ? crafting_inv.charges_of( item_id,
-                                 std::numeric_limits<long>::max(), filter ) : crafting_inv.amount_of( item_id, false,
+                                 std::numeric_limits<long>::max(), filter ) : (long)crafting_inv.amount_of( item_id, false,
                                          std::numeric_limits<long>::max(), filter );
                 text = item::count_by_charges( item_id ) ?
                        string_format( _( "%s (%d of %ld)" ), item::nname( item_id ), component.count * batch, qty ) :
